+++ Intent.Modules.NodeJS.Services.CRUD/CrudStrategies/CreateImplementationStrategy.cs
@@ -50,7 +50,7 @@ namespace Intent.Modules.NodeJS.Services.CRUD.CrudStrategies
                 return false;
             }
 
-            if (_template.TryGetTypeName(RepositoryTemplate.TemplateId, targetEntity, out _repository))
+            if (!_template.TryGetTypeName(RepositoryTemplate.TemplateId, targetEntity, out _repository))
             {
                 return false;
             }