+++ Intent.Modules.NodeJS.Services.CRUD/CrudStrategies/DeleteImplementationStrategy.cs
@@ -56,7 +56,7 @@ namespace Intent.Modules.NodeJS.Services.CRUD.CrudStrategies
                 return false;
             }
 
-            if (_template.TryGetTypeName(RepositoryTemplate.TemplateId, targetEntity, out _repository))
+            if (!_template.TryGetTypeName(RepositoryTemplate.TemplateId, targetEntity, out _repository))
             {
                 return false;
             }