+++ Intent.Modules.TypeORM.Entities/Templates/TemplateExtensions.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using Intent.Modules.Common.Templates;
+using Intent.Modules.TypeORM.Entities.Templates.BasicAuditingSubscriber;
 using Intent.Modules.TypeORM.Entities.Templates.Datasource;
 using Intent.Modules.TypeORM.Entities.Templates.OrmConfig;
 using Intent.Modules.TypeORM.Entities.Templates.Repository;
@@ -15,6 +16,10 @@ namespace Intent.Modules.TypeORM.Entities.Templates
 {
     public static class TemplateExtensions
     {
+        public static string GetBasicAuditingSubscriberName<T>(this IntentTemplateBase<T> template)
+        {
+            return template.GetTypeName(BasicAuditingSubscriberTemplate.TemplateId);
+        }
         public static string GetDatasourceName<T>(this IntentTemplateBase<T> template)
         {
             return template.GetTypeName(DatasourceTemplate.TemplateId);